--- src/components/propertylist/propertylist.module.css
@@ -1,33 +1,0 @@
-@import "../../core/styles/index.css";
-
-
-.verifyDeleteContainer {
-    position: fixed;
-    height: 160px;
-    width: 300px;
-    text-align: center;
-    background: #f1f5f8;
-    top: 50%;
-    left: 50%;
-    margin-top: -80px;
-    margin-left: -150px;
-    border: 1px solid #404446;
-    color: #404446;
-    -webkit-box-shadow: 0 0 5px 2px #eee;
-    -moz-box-shadow: 0 0 5px 2px #eee;
-    box-shadow: 0 0 5px 2px #eee;
-    border-radius: 3px;
-}
-
-.buttonContainer {
-    position: absolute;
-    text-align: center;
-    width: 100%;
-    bottom: 10px;
-}
-
-.verifyDeleteText {
-    position: absolute;
-    font-size: 16px;
-    margin: 10px;
-}
