+++ src/views/Main/Schemas/Components/PropertyList.js
@@ -1,11 +1,11 @@
 import React, { PropTypes } from 'react';
 import { EntityDataModelApi } from 'loom-data';
 import { Property } from './Property';
-import StringConsts from '../../utils/Consts/StringConsts';
-import PermissionsConsts from '../../utils/Consts/PermissionConsts';
-import { NameNamespaceAutosuggest } from '../namespaceautosuggest/NameNamespaceAutosuggest';
-import Utils from '../../utils/Utils';
-import styles from './propertylist.module.css';
+import StringConsts from '../../../../utils/Consts/StringConsts';
+import PermissionsConsts from '../../../../utils/Consts/PermissionConsts';
+import { NameNamespaceAutosuggest } from '../../../../components/namespaceautosuggest/NameNamespaceAutosuggest';
+import Utils from '../../../../utils/Utils';
+import styles from '../styles.module.css';
 
 export class PropertyList extends React.Component {
   static propTypes = {
